Constraints

Attributes

PolyJuMP.MomentsAttribute — Type

MomentsAttribute(N)
MomentsAttribute()

A constraint attribute for the vector of moments corresponding to the constraint that a polynomial is zero in the full space in result N. If the polynomial is constrained to be zero in an algebraic set, it is the moments for the constraint once it is rewritten into an constraint on the full space. If N is omitted, it is 1 by default.

Examples

Consider the following program:

@variable(model, α)
@variable(model, β ≤ 1)

using DynamicPolynomials
@polyvar x y
cref = @constraint(model, α * x - β * y == 0, domain = @set x == y)

The constraint is equivalent to

@constraint(model, (α - β) * y == 0)

for which the dual is the 1-element vector with the moment of y of value -1. This is the result of moments(cref). However, the dual of cref obtained by dual(cref) is the 2-elements vector with both the moments of x and y of value -1.

MultivariateMoments.moments — Method

moments(cref::JuMP.ConstraintRef)

Return the MomentsAttribute of cref.

SumOfSquares.GramMatrix — Type

struct GramMatrix{T, B, U, MT <: AbstractMatrix{T}} <: AbstractGramMatrix{T, B, U}
    Q::MT
    basis::B
end

Gram matrix $x^\top Q x$ where Q is a symmetric matrix indexed by the vector of polynomials of the basis basis.

SumOfSquares.GramMatrixAttribute — Type

GramMatrixAttribute(result_index)
GramMatrixAttribute()

A constraint attribute for the GramMatrix of a constraint, that is, the positive semidefinite matrix Q indexed by the monomials in the vector X such that $X^\top Q X$ is the sum-of-squares certificate of the constraint.

SumOfSquares.gram_matrix — Function

gram_matrix(cref::JuMP.ConstraintRef)

Return the GramMatrixAttribute of cref.

SumOfSquares.gram_operate — Function

gram_operate(::typeof(+), p::GramMatrix, q::GramMatrix)

Computes the Gram matrix equal to the sum between p and q. On the opposite, p + q gives a polynomial equal to p + q. The polynomial p + q can also be obtained by polynomial(gram_operate(+, p, q)).

gram_operate(/, p::GramMatrix, α)

Computes the Gram matrix equal to p / α. On the opposite, p / α gives a polynomial equal to p / α. The polynomial p / α can also be obtained by polynomial(gram_operate(/, p, α)).

SumOfSquares.MomentMatrixAttribute — Type

MomentMatrixAttribute(N)
MomentMatrixAttribute()

A constraint attribute fot the MomentMatrix of a constraint.

MultivariateMoments.moment_matrix — Function

moment_matrix(cref::JuMP.ConstraintRef)

Return the MomentMatrixAttribute of cref.

SumOfSquares.CertificateBasis — Type

struct CertificateBasis <: MOI.AbstractConstraintAttribute end

A constraint attribute for the basis indexing the GramMatrixAttribute and MomentMatrixAttribute certificates.

SumOfSquares.certificate_basis — Function

certificate_basis(cref::JuMP.ConstraintRef)

Return the CertificateBasis of cref.

SumOfSquares.certificate_monomials — Function

certificate_monomials(cref::JuMP.ConstraintRef)

Return the monomials of certificate_basis. If the basis if not MultivariateBases.AbstractMonomialBasis, an error is thrown.

SumOfSquares.LagrangianMultipliers — Type

LagrangianMultipliers(N)
LagrangianMultipliers()

A constraint attribute fot the LagrangianMultipliers associated to the inequalities of the domain of a constraint. There is one multiplier per inequality and no multiplier for equalities as the equalities are handled by reducing the polynomials over the ideal they generate instead of explicitely creating multipliers.

SumOfSquares.lagrangian_multipliers — Function

lagrangian_multipliers(cref::JuMP.ConstraintRef)

Return the LagrangianMultipliers of cref.

SumOfSquares.SOSDecomposition — Type

struct SOSDecomposition{T, PT}

Represents a Sum-of-Squares decomposition without domain.

SumOfSquares.SOSDecompositionWithDomain — Type

struct SOSDecompositionWithDomain{T, PT, S}

Represents a Sum-of-Squares decomposition on a basic semi-algebraic domain.

SumOfSquares.SOSDecompositionAttribute — Type

struct SOSDecompositionAttribute
    ranktol::Real
    dec::MultivariateMoments.LowRankLDLTAlgorithm
    result_index::Int
end

A constraint attribute for the SOSDecomposition of a constraint. By default, it is computed using SOSDecomposition(gram, ranktol, dec) where gram is the value of the GramMatrixAttribute.

SumOfSquares.sos_decomposition — Function

sos_decomposition(cref::JuMP.ConstraintRef)

Return the SOSDecompositionAttribute of cref.

sos_decomposition(cref::JuMP.ConstraintRef, K<:AbstractBasicSemialgebraicSet)

Return representation in the quadraic module associated with K.

SAGE decomposition attribute:

PolyJuMP.SAGE.Decomposition — Type

struct Decomposition{T, PT}

Represents a SAGE decomposition.

PolyJuMP.SAGE.DecompositionAttribute — Type

struct DecompositionAttribute{T} <: MOI.AbstractConstraintAttribute
    tol::T
end

A constraint attribute for the Decomposition of a constraint.